# tagooledavid-folder/week12/Week12 - CapstoneApiweek12/api-client/models/StudentApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Student } from './Student';
export type StudentApiResponse = {
    status?: string;
    message?: string;
    data?: Student;
};