# tagooledavid-folder/week12/Week12 - CapstoneApiweek12/api-client/models/StudentListApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Student } from './Student';
export type StudentListApiResponse = {
    status?: string;
    message?: string;
    data?: Array<Student>;
};