# tagooledavid-folder/week13/Week13 - CapstoneApiweek13/api-client/models/NewStudent.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type NewStudent = {
    name: string;
    regNumber: string;
    gpa: number;
};